requirements.txt
pdfminer.six==20250506
pdf2image==1.17.0
pillow==11.3.0
pytesseract==0.3.13